# specs/001-curator-quiz-frontend/contracts/advanced-features.openapi.yaml
openapi: 3.1.0
info:
  title: Duck Emporium Advanced Features API
  version: 0.1.0
  description: >-
    Contracts for stories 06-09: curator add duck, Duck of the Day, personality
    quiz, and static web frontend entrypoint.
servers:
  - url: http://localhost:3000
paths:
...
  /admin/ducks:
    post:
      summary: Add a new duck to catalog (curator only)
      operationId: addDuck
      parameters:
        - name: x-admin-password
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDuckRequest'
      responses:
        '201':
          description: Duck created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDuckSuccess'
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing/invalid admin password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate duck name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    AddDuckRequest:
      type: object
      required:
        - name
        - category
        - price
        - tagline
        - description
        - personalityTraits
        - initialStock
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        tagline:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        personalityTraits:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        initialStock:
          type: integer
          minimum: 0
    AddDuckSuccess:
      type: object
      required:
        - duck
      properties:
        duck:
          $ref: '#/components/schemas/Duck'
    Duck:
      type: object
      required:
        - id
        - name
        - category
        - price
        - tagline
        - description
        - personalityTraits
        - stock
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        tagline:
          type: string
        description:
          type: string
        personalityTraits:
          type: array
          items:
            type: string
        stock:
          type: integer
          minimum: 0
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string